# Write unary content with single Stream.WriteAsync