Large performance improvement in unary RPS.

https://github.com/dotnet/runtime/issues/35184#issuecomment-626304981

Before
```
Successfully processed 100373; RPS 27299; Errors 0; Last elapsed 3.9209ms
Successfully processed 127896; RPS 27496; Errors 0; Last elapsed 3.6429ms
Successfully processed 155281; RPS 27374; Errors 0; Last elapsed 3.6593ms
Successfully processed 182801; RPS 27520; Errors 0; Last elapsed 3.546ms
Successfully processed 209917; RPS 27112; Errors 0; Last elapsed 3.6468ms
```
After
```
Successfully processed 228755; RPS 52340; Errors 0; Last elapsed 2.1993ms
Successfully processed 282848; RPS 53993; Errors 0; Last elapsed 2.1865ms
Successfully processed 336652; RPS 53802; Errors 0; Last elapsed 1.5827ms
Successfully processed 390310; RPS 53628; Errors 0; Last elapsed 0.9105ms
Successfully processed 443199; RPS 52889; Errors 0; Last elapsed 2.2438ms
```

When Google.Protobuf supports writing to `IBufferWriter<byte>` this can be improved to write message data directly to the buffer array.

@stephentoub